backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
torch
torchvision
opencv-python-headless
numpy
pillow
kagglehub
